# Item.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace demo
{
    public class Item
    {
        public int Id{get;set;}
        public string description{get;set;}
    }
}